Fix initial camera quality
When a call started and the VideoConstrainer was created the video track
was assumed to be already in high quality, so when the VideoConstrainer
was asked to apply again the high quality constraints it just ignored
the request. However, in some cases the browser does not initialize the
camera in high quality, so the camera was kept in a lower quality for
as long as the high quality was kept being requested.

Moreover, even if the browser initialized the camera with high quality
that quality might not be exactly the same as the one set when applying
the high quality constraints, so even in those cases the constraints
should be applied anyway for consistency.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>