Add indicator of unread messages while in call
When in a call and the sidebar is collapsed, whenever new unread
messages arrive, a bubble will be displayed on top of the button that
opens the sidebar.

If a mention exists, the bubble will appear highlighted.

Whenever a new unread message or new mention arrives while in a call
with the sidebar closed, a notification will appear once.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>
Co-authored-by: Joas Schilling <213943+nickvergessen@users.noreply.github.com>