Fix switching devices in Firefox
If a device is currently active and a different one is requested Firefox
returns the currently active device. To overcome this all the tracks of
the same device kind as the requested one are stopped before performing
the request.

However, a device is also "active" as long as any other device returned
in the same "getUserMedia" request is active. Both audio and video are
requested at once when a call is started, so during calls it was needed
to stop both devices first to be able to switch just one of them.

Moreover, since Firefox 88 there is a grace period for requesting again
the media permissions to the user once a track stops. During that grace
period the device is also "active" and it is not possible to switch to
another device even if the previous one was requested on its own,
without any other device. In this case the permissions have to be
manually revoked to be able to switch to another device before the grace
period ended.

To solve all that now the ID of the requested device is marked as a
required constraint instead of an optional one, which forces Firefox to
honour the requested device and return it instead of the active one.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>