Unify STUN and TURN server URLs
The "stunservers" parameter was a list of several elements each one with
a "url" parameter, which was a string. Now the "stunservers" parameter
is a list of several elements (although in practice there will be just
one) each one with a "urls" parameter, which is a list of strings.

The "turnservers" parameter was a list of several elements each one with
a "url" parameter and a "urls" parameter, which were both a list of a
single string. Now the "turnservers" parameter is a list of several
elements each one with a "urls" parameter, which is a list of strings.
Each element of "turnservers" contain too "username" and "credential"
parameters that apply to all the elements in the "urls" parameter.

The format resembles the RTCIceServer format, so the returned values can
be directly used by the WebUI like done until now. Mobile clients will
need to be adjusted, though.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>