Handle track changes when call was started without audio or video
When the call was started without audio nor video (either explicitly or
because there was an error when getting the media) the media device
changes were not listened to, so further device changes were ignored and
"localTrackReplaced" events were not emitted.

Moreover, "localTrackReplaced" events were handled by Peer objects to
replace or add tracks as needed, but if a call is started without media
some Peer objects will be missing (the ownPeer object if the HPB is
used, Peer objects for any other participant that is not sending media
when the HPB is not used). Therefore now a forced reconnection is
triggered when needed to establish the missing sender connections.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>