Do not show an error if explicitly joining a call without media
"TypeError" is thrown by "getUserMedia" when requesting media without
audio nor video. In that case no error message should be shown, as the
devices would have been explicitly disabled by the user.

That could also happen when audio was disabled and video was enabled but
failed. If the initial "getUserMedia" call done when starting a call
failed and video was requested it was retried again without video in all
cases. Now it is only retried if audio is also enabled, as otherwise it
would end requesting no audio and no video (and no error message would
be shown, as a "TypeError" would be thrown).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>