Do not add tracks for receiver only peers
When the HPB is used the peer connections for the other participants
only receive the media, but do not send any on them (the data is sent on
a single dedicated sending peer). Therefore there is no need to add
tracks to those peers, nor listen to changes in the local tracks to
update them.

The "broadcaster" property was already set before only on sending peers,
so nothing is changed in that regard either.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>