Fix call flags not updated when disabling or reenabling devices
When a media device is disabled during a call its track is replaced with
a null track in the peers, and if a media device is reenabled the null
track is replaced with the new track. In both cases neither a
renegotiation nor a reconnection is needed. Therefore the call flags
have to be explicitly updated.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>