Include "publishingPermissions" in participant messages of signaling
This will allow the clients to react to "publishingPermissions" changes
directly from the signaling messages, without needing to fetch again the
participants (although they may need to fetch them again nevertheless
for UI updates).

The internal signaling server also needs to listen to changes on the
property to be able to know when to send the message (the external
signaling server already listened to the changes to be able to update
the permission flags internally).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>