Listen to local media events only once when joining a call
Until now the "localMediaStarted" and "localMediaError" were emitted
only when the local participant joined a call. Therefore it was possible
to listen to all the events when waiting for the media to be ready
before joining a call.

However, in order to be able to call "startLocalVideo()" again during a
call (for example, if publishing permissions were granted) the events
must be listened to only once when waiting for the media to be started
to then join a call.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>