Keep track of whether the local media is active or not
Until now it was only possible to know if there were active local
streams, but not whether the local media as a whole was active or not.
That is, if new streams would be automatically started if media devices
were switched. This will be needed to properly react to changes in the
publishing permissions to ensure that even if there are no streams no
streams will be started at a later point when there are no publishing
permissions.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>