Disable audio and video buttons also if publishing is not allowed
The audio and video buttons were enabled and disabled based on whether
audio and video was available. This roughly matched the publishing
permissions, but after a permission change there could be a small lapse
of time in which audio and video was available but publishing was not
allowed, which caused a mismatch between the button being enabled and
the tooltip stating that enabling audio/video was not allowed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>