Don't trigger leave event when deleting conversation
When deleting a conversation, the route change would implicitly trigger
a leave request which ran in parallel to the conversation deletion one,
causing random issues in some systems like the latter returning a 404
instead of proceeding with the deletion.

This fix removes the route change on deletion. The new behavior will
implicitly redirect to the "not-found" page since the session does not
exist any more. The latter behavior is now aligned with the "Leave
conversation" action which also redirects to "not-found" page.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>