Fall back get participant in case session is gone
In some concurrent request scenarios, the session of a participant might
expire just after its id got retrieved. To avoid unexpected behavior, if
the participant could not be found by session id, fall back to getting
it by user id.

Fixes issue where a room can sometimes not be deleted from the UI due to
a concurrent request that removes the session id.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>