Ignore current participant if removed concurrently
When retrieving the conversation list for a user, there is a
time window between retrieving the list and the participant information.
If in that time window, a concurrent request removes the user from the
room, or even the room got deleted, this could cause a
ParticipantNotFoundException.

This fix prevents the latter exception to make the whole call return a
404. Instead, it skips the obsoleted entry and carries on with the list.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>