Reset stats before the analysis is started
The stats needs to be reset when a peer connection starts. The stats
from the previous peer connection are unrelated to the new one, and that
will cause a wrong analysis until the stats are filled again with values
from the new peer connection.

Similarly, the stats should be reset too when a peer connection is
restarted. In this case the reported stats continue from the last
values. However, during the reconnection the stats will not be updated,
so the timestamps will suddenly increase once the connection is ready
again, which could cause a wrong analysis.

Due to all that now the stats are reset before the analysis is started,
as the analysis is started only when the peer connection is started or
restarted.

Note that right now the sender connections with the HPB are never
restarted; a new peer connection is always established. Nevertheless,
this approach should be more "future proof".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>