Expose connection state in LocalCallParticipantModel
When the MCU is used there will be a sender peer connection (the so
called "own peer"). The ICE connection state of that peer connection is
now exposed in LocalCallParticipantModel in the same way as in
CallParticipantModel. However, unlike the CallParticipantModel, the
LocalCallParticipantModel is always set and the peer object set or unset
on it as needed; if there is no peer object set the connection state
will not be set.

When the MCU is not used there is no sender peer connection (nor any
other peer associated to the LocalCallParticipantModel), so in that case
the connection state attribute will not be set either.

Besides that some logs were also added for the ICE connection state
changes of the own peer, similar to the logs already present for the
peer connection of the other participants.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>