Fix concurrency when cancelling lookForNewMessages
When switching between regular mode and call mode, the two MessagesList
view instances are concurrently disturbing each other's
lookForNewMessages call.

To fix this, we now introduce a requestId for those requests to which we
pass in the chatIdentifier, extended by a generated view id. This brings
back the old behavior for now.

In the future we should look into keeping the polling request and
sharing it between the two Vues, which requires some refactoring of the
polling logic and timeouts.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>