Do not replace tracks when a connection has not been established yet
If a connection has not started yet tracks can not be replaced, as
Firefox will get "stuck" and not replace the tracks even if tried later
again once connected.

To address that now the tracks are replaced only when the connection is
not in the "new" state (even if the connection is in the "checking"
state replacing the track seems to work fine). If a track is tried to be
replaced in the "new" state the action is enqueued and executed once the
connection is started.

Besides that as the replace track actions are now enqueued this also
prevents replacing a track before the previous replacement has finished.
This should not happen, but if it did it could have caused issues due to
the track to be replaced not having been set yet.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>