Do not process pending replace tracks again if they are being processed
If "processPendingReplaceTracks" was called while a processing was
on-going (for example, on quick connection state changes) this would
cause another processing in parallel, which could have unexpected
results.

To prevent that now there is at most a single processing on-going at any
time. The method also returns a promise now, which will make possible to
enqueue further actions after the processing has finished even if the
processing was actually started in a different call.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>