Emit sidebar events on Nextcloud bus
The viewer expect the "files:sidebar:opened" and "files:sidebar:closed"
events to be emitted on Nextcloud bus to adjust its width based on the
sidebar.

Both events could be emitted by directly listening to AppSidebar events
on RightSidebar. However, for now only the "closed" event is listened;
the "opened" event does not seem to work reliably (it seems that
sometimes it is emitted before the sidebar has finished opening, so the
viewer ends overlapping the sidebar due to getting its width when it is
not fully opened), so for now the custom detection of when the sidebar
has finished opening is still used.

In Talk the sidebar events are currently needed only by the viewer, and
when the viewer is opened the sidebar will be opened only by the viewer
through "OCA.Talk.Sidebar.open()", so not emitting the sidebar events
directly on the RightSidebar component should not be a problem.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>