Add logs for connection quality warnings
The logs are printed when the PeerConnectionAnalyzer changes to a state
that causes a connection quality warning to be shown (very bad quality
or not transmitted data).

Currently the connection quality is analyzed only when the HPB is used
and only for the sender participant, so there will be at most a single
PeerConnectionAnalyzer. Due to this, for simplicity, for now the stats
are logged without any participant identifier.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>