Disable simulcast for screenshares for now
Simulcast does not currently work for screenshares neither in Firefox
nor in Chromium. Moreover, if a screen is shared in Chromium and
simulcast is enabled for the peer connection of the screen Janus logs
will be spammed with "[WARN] Incoming RTCP SR SSRC (XXXXXXXXX) does not
match the expected one (YYYYYYYYYY) video=1" messages. Therefore, for
now, simulcast is disabled for screen peers.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>