Fix initial camera quality with Chromium
Due to a bug in Chromium it is not possible to increase the resolution
of a track once it has been cloned.

When the video was initially requested by Talk no resolution was
specified, so Chromium returned a video track with a ~640x480
resolution, no matter the maximum resolution of the camera. As the track
was immediately cloned, due to the bug, the maximum resolution of the
track was limited to ~640x480.

Now, if Chromium or a derived browser is used, the video is requested
with a loose constraint for a high resolution, so if the camera does not
have such resolution it will still return the highest resolution
available without failing. If the track is then cloned then it will be
possible to decrease and increase again the track resolution as desired
instead of being locked to a maximum resolution of ~640x480.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>