Use visual last read message id in scroll handler
The scroll handler checks if the unread marker was seen before doing
updates after scrolling down. After updating the read marker in the
store, its value becomes different than the visual one. Because of this,
the check for seen was done on the wrong element and would always be
false.

This fix makes the handler use the visual marker instead of the store
one when checking for update. This makes sure that the marker will
continue to be updated when scrolling down, but will only allow
increasing the message id past the non-visual one to prevent bringing
the marker back up.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>