Fix read marker and counter issues with pending pages
Whenever there are several pages of unread messages, the
"lookForNewMessages()" call will not return all the results at once.

In case there are results left to load, the read marker is not reset by
the false when "isSticky" (aka scrolled to bottom) happens.

When loading new messages, we now only increase the conversation counter
when the latter was out of sync with the current messages list.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>