Integrated tflite selfieSegmentation. Following problems need to be solved:
Stopping a call does not stop the tflite processes. Restarting
the call is possible but starts new instances of the segmentation
processes which results in rising cpu load with every newly started
call.

The segmentation pipeline is chosen by availability. That results in bad
performance if webgl is available but 3D acceleration is turned off or
your graphics card is weak. Canvas 2d offers much better performance in
these cases. The easiest solution would be to let the user choose the
pipeline.

Signed-off-by: Immanuel Pasanec <immanuel.pasanec@compaso.de>