Fix duplicated registration of audio encoder
The audio encoder is initialized when the AudioRecorder component is
mounted, and the store keeps track of whether the audio encoder was
already initialized to not do it again.

However, the audio encoder was also unconditionally registered in the
main components of the main and sidebar Talk UIs, which caused a
duplicated registration when the audio encoder was initialized (as the
store did not "know" that it was already registered).

Due to all this the unconditional registration is removed (which also
avoids registering the audio encoder when it will not be needed, like
when the current user is a guest without upload permissions).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>