Fix capping when decreasing the minimum value in constraints
The minimum value of the constraints was wrongly capped using "min"
instead of "max", so in practice the first time that the value was
decreased it was already set to the minimum capped value. Moreover, if
the value could not be set the minimum value could be decreased in the
next iteration even below the capped value.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>