Fix connection quality warning still shown after media is stopped
When the audio/video or the screen peers were cleared the connection
quality data was not reset. Due to this the connection quality warning
would be kept shown if, for example, the screen share was stopped while
the warning was being shown.

For simplicity and consistency the connection quality data is reset too
when a different peer connection is set (not just when it is cleared),
although this is not really needed as the previous connection quality
was already cleared in that case further in the code path.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>