Fix "stopped speaking" events still emitted once started to speak again
The "stopped speaking" events emitted by the audio monitor are not
immediately relayed, but once a grace period of one second has elapsed.
However, when the user starts to speak again the timeout was not
cleared, so the "stopped speaking" event was emitted even if the user
started to speak again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>