Fix null stream given to RTCPeerConnection.addTrack
If a call is started and no audio nor video device is selected there
will be no stream in the local media. If a device is then selected a new
MediaStream will be created to add the track to it. However, the event
did not include the stream that the track was added to but a null
object, so when the event was handled calling
"RTCPeerConnection.addTrack()" failed due to being given a null stream.
Now the event includes the stream that the track was added to, which
will be either the already existing stream or the one just created.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>