Extract common emitter code to mixin
The EmitterMixin is not a Vue mixin but a raw JavaScript mixin, so it is
stored in "src/utils" rather than in "mixins".

The emitter code was almost the same in all classes, except that some
classes allowed triggering events without parameters and others did not.
The mixin uses the most generic code, that is, the one that also allows
triggering events without parameters.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>