Add SpeakingMonitor node to emit "speaking" events from an audio track
The node replaces the audio monitors in the LocalMedia object. For
simplicity, and as this is just a temporal step in the refactoring, it
is assumed that there will be just a single local stream, so only one
SpeakingMonitor node is used now (instead of having one audio monitor
for each stream like before, although in most cases there was just one
stream at any time).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>