Add base classes for media pipeline
A media pipeline is a directed graph with nodes that act as a source, a
sink or both. A source node provides one or more output slots and a sink
node provides one or more input slots. The output slot of a source can
be connected to several input slots of a sink, while an input slots of a
sink can be connected only to a single output slot of a source.

This commit introduces just the base classes used to route tracks; these
base classes need to be extended to provide nodes with specific
behaviours.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>