Do not explicitly disable audio and video on changed streams
When a stream was changed the audio and video was explicitly disabled
again even if it was already disabled. This was needed in the past
because updated tracks always started enabled, so they had to be
explicitly disabled through the model. However, since 82ba579624 the
updated tracks are automatically disabled if needed after starting, so
they are already consistent with the state in the model and thus it is
no longer needed to explicitly disable them again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>