Fix enabled state not updated in device checker
The enabled/disabled state of audio and video was set when the device
checker was mounted. However, the device checker is mounted just once
when Talk is loaded and then shown and hidden as needed, so the enabled
state did not reflect the current value when starting a call again, only
what was last set in the device checker.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>