Show background blur in the device checker
The video stream got from MediaDevicesManager is now passed to a
VirtualBackground node, and its output is converted again to a stream
like the one used before with a TrackToStream node. In the future it
might be better to use a full media pipeline, but for now this is good
enough.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>