Fix grid when the number of videos is below the limit
When "videosCap" is set it was always applied, no matter the actual
number of videos. Due to this, even if there was just a single remote
video, as the available size was calculated based on the limit rather
than on the actual number of videos a mostly empty grid was shown.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>