Ignore preference for H.264 when simulcast is used
When "prefer_h264" is set the SDP is modified to place H.264 as the
preferred codec. However, this change in the SDP clashes with the
changes done to enable simulcast with Chromium and sometimes causes
creating the offer to fail. Given that preferring H.264 has no effect
when the HPB is used (as only VP8 is supported in that case) and that
simulcast can be enabled only with the HPB now "prefer_h264" is ignored
when simulcast is used.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>