Remove setting to prefer H.264
Back in the day the setting to prefer H.264 was added as a technology
preview; it was not a finished feature, but a preview of a feature that
could be added in the future. However, H.264 does not work when the HPB
is used, it overrides the preference order already decided by the
browser and it does so by altering the SDP, which could lead to problems
if its format changes. Due to all that preferring H.264 is no longer
available.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>