Show message about problems in the connection with another participant
Until now when a participant was not connected a loading spinner was
shown on that participant. Now an explicit message is shown when there
are connection problems (which may heal by themselves) and when the
connection is tried to be established again to solve the problems
(either because the connection actually failed, or because the problems
did not heal by themselves after several seconds, even if the connection
has not failed yet).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>