Disable virtual background if TFLite model could not be loaded
Until now it was checked if the blur background was supported based on
whether Web Assembly was available in the browser and whether the
browser supported the needed canvas filter. This could be done
statically, without having a VirtualBackground object. However, to blur
the background once the VirtualBackground object is created it is also
needed to download and compile a TFLite model; this could fail, for
example, if the web server is not properly configured to send .wasm
files. Now if virtual background is supported but the model failed to
load it behaves as if virtual background is not supported.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>