Add setup check for background blur
A new section is added to Talk administration settings for web server
checks.

The check creates a JitsiStreamBackgroundEffect object to ensure that
the background blur files could be initialized, rather than implementing
its own logic to verify the .wasm and .tflite files. Due to this, the
browser running the check needs to support and have WebAssembly enabed,
although it does not need to support the background blur itself (like
Safari, which does not implement the needed canvas filters, but can
anyway be used to do the check).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>