Update the output stream of the background blur instead of changing it
When the constraints of the input stream were changed the background
blur was stopped and started again to apply the new width, height and
frame rate to the output. However, changing a stream (any stream, it is
unrelated to the background blur) causes a flickering (one or more black
frames in between), so every time that the sent video quality was
adjusted the local video flickered.

Now, instead of reseting the background blur, the internal elements are
updated and adjusted to the new constraints of the input, but the same
output stream is kept. This avoids the stream change and thus the
flickering. Note, however, that the video will temporary freeze instead
while the input stream is being re-rendered, although this should be
less annoying than the flickering.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>